FLIPKART-GRID-main/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
bidict==0.23.1
blinker==1.8.2
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
easyocr==1.7.2
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonify==0.5
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.1
ninja==1.11.1.1
numpy==1.26.4
openai==1.52.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.2
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==75.1.0
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
